Add health.npi_registry_search()
New function queries against the search url instead of the API. This was tested generally, but not as written so there are notes and todo items that can be resolved relatively fast (IDE is down until Tuesday/Wednesday). get_providers() needs to be re-thought to drop df already appended to dfs since this will return duplicates. Generally it could use a solid refactor as it violates DRY. Tried to keep it scoped to working under existing get_providers, but able to be made extensible to more general search params (probably over-engineered). Hope is they will add exactMatch param to API so this can just use one url/pattern.